+++ lib/src/behavior/scroll_behavior.dart
@@ -22,5 +22,23 @@ class ERScrollBehavior extends ScrollBehavior {
   }
 
   @override
+  Widget buildScrollbar(BuildContext context, Widget child, ScrollableDetails details) {
+    switch (getPlatform(context)) {
+      case TargetPlatform.linux:
+      case TargetPlatform.macOS:
+      case TargetPlatform.windows:
+        assert(details.controller != null);
+        return Scrollbar(
+          controller: details.controller,
+          child: child,
+        );
+      case TargetPlatform.android:
+      case TargetPlatform.fuchsia:
+      case TargetPlatform.iOS:
+        return child;
+    }
+  }
+
+  @override
   Set<PointerDeviceKind> get dragDevices => _kDragDevices;
 }